move VanillaMenuBuilder and add support for List2 contextual menu
